Change the frameworks folder structure
