--- App/Core/Exceptions/EnvFileNotFoundException.php
@@ -1,15 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Core\Exceptions;
-
-use Exception;
-
-class EnvFileNotFoundException extends Exception
-{
-    public function __construct(string $message)
-    {
-        parent::__construct($message);
-    }
-}
