--- App/Core/Http/HttpMethod.php
@@ -1,14 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Core\Http;
-
-enum HttpMethod: string
-{
-    case Get = 'GET';
-    case Post = 'POST';
-    case Patch = 'PATCH';
-    case Delete = 'DELETE';
-
-}
