--- App/Routes/Routes.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Routes;
-
-enum Routes: string
-{
-    case Homepage = '/';
-    case AppError = '/chyba-aplikace';
-
-}
