+++ Core/Exceptions/DbQueryException.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Core\Exceptions;
+
+use Exception;
+
+class DbQueryException extends Exception
+{
+    public function __construct(string $message)
+    {
+        parent::__construct($message);
+    }
+}
