+++ Core/Model/Model.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Core\Model;
+
+use PDO;
+
+class Model
+{
+    public function __construct(protected PDO $connection)
+    {
+    }
+
+    public function __destruct()
+    {
+        unset($this->connection);
+    }
+}
