+++ Core/Utils/Sanitize.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Core\Utils;
+
+// Jednotuchá třída na základní ošetření vstupů
+class Sanitize
+{
+    public static function sanitize(string|array $data): string|array
+    {
+        if (is_array($data)) {
+            $processedArray = [];
+
+            foreach ($data as $key => $value) {
+                $processedArray[$key] = trim(htmlspecialchars($value));
+            }
+
+            return $processedArray;
+        }
+
+        return trim(htmlspecialchars($data));
+    }
+}
