+++ src/Routes/Routes.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace src\Routes;
+
+enum Routes: string
+{
+    case Homepage = '/';
+    case AppError = '/chyba-aplikace';
+
+}
